--- app/platform.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-void platform_delay_us(uint32_t delay_us);
