+++ platforms/stm32l4xx/config/stm32l4xx_hal_conf.h
@@ -49,7 +49,7 @@
 #include "stm32l4xx_hal_tim.h"
 #include "stm32l4xx_hal_uart.h"
 
-#ifdef  USE_FULL_ASSERT
+#ifdef USE_FULL_ASSERT
 	#define assert_param(expr) ((expr) ? (void)0U : assert_failed((uint8_t *)__FILE__, __LINE__))
 	void assert_failed(uint8_t* file, uint32_t line);
 #else
