+++ targets/target.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "errors.h"
+
+/* This file is located in targets/$(TARGET)/target_conf.h */
+#include "target_conf.h"
+
+err_t target_init(void);
