+++ targets/xildebug_big/build.mk
@@ -1,0 +1,4 @@
+PLATFORM := stm32l4xx
+
+TARGET_SRCS += \
+	targets/$(TARGET)/target.c \
