+++ targets/xildebug_big/target.c
@@ -1,0 +1,6 @@
+#include "platform/platform.h"
+
+err_t target_init(void)
+{
+	return platform_init();
+}
